requirements.txt
PyMuPDF==1.26.3
sentence-transformers==5.0.0
nltk==3.9.1
faiss-cpu==1.11.0
scikit-learn==1.4.2
tqdm==4.66.2
fastapi==0.116.1
uvicorn==0.35.0